refactor: Replace placeholder image
